Split out `sut` (system-under-test) module
